@@ -26,18 +26,17 @@
     "object-assign": "4.1.1",
     "promise": "8.0.1",
     "react": "^15.6.1",
-    "react-anything-sortable": "^1.7.4",
-    "react-beautiful-dnd": "^7.1.3",
     "react-dom": "^15.6.1",
     "react-particles-js": "^1.5.0",
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
+    "react-smooth-dnd": "^0.4.9",
     "redux": "^3.7.2",
     "redux-actions": "^2.2.1",
     "redux-devtools-extension": "^2.13.2",
     "redux-saga": "^0.15.6",
     "sw-precache-webpack-plugin": "0.11.4",
-    "swiper": "^4.2.6",
+    "swiper": "^4.3.3",
     "whatwg-fetch": "2.0.3"
   },
   "scripts": {
@@ -87,35 +86,35 @@
     "extends": "react-app"
   },
   "devDependencies": {
-    "babel-core": "6.25.0",
+    "babel-core": "^6.26.3",
     "babel-eslint": "7.2.3",
     "babel-jest": "20.0.3",
     "babel-loader": "7.1.1",
     "babel-plugin-import": "^1.5.0",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-polyfill": "^6.26.0",
-    "babel-preset-react-app": "^3.0.2",
+    "babel-preset-react-app": "^3.1.1",
     "babel-runtime": "6.26.0",
     "css-loader": "0.28.4",
-    "eslint": "4.4.1",
+    "eslint": "^4.19.1",
     "eslint-config-react-app": "^2.0.0",
     "eslint-loader": "1.9.0",
     "eslint-plugin-flowtype": "2.35.0",
-    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-import": "^2.12.0",
     "eslint-plugin-jsx-a11y": "5.1.1",
     "eslint-plugin-react": "7.1.0",
     "extract-text-webpack-plugin": "3.0.0",
     "less": "^2.7.2",
     "less-loader": "^4.0.5",
     "mockjs": "^1.0.1-beta3",
+    "postcss-flexbugs-fixes": "3.2.0",
+    "postcss-loader": "2.0.6",
+    "react-dev-utils": "^4.0.1",
     "style-loader": "0.18.2",
     "url-loader": "0.5.9",
     "webpack": "3.5.1",
     "webpack-dev-server": "2.7.1",
     "webpack-manifest-plugin": "1.2.1",
-    "react-dev-utils": "^4.0.1",
-    "postcss-flexbugs-fixes": "3.2.0",
-    "postcss-loader": "2.0.6",
     "yarn": "^1.7.0"
   }
 }
