@@ -1,5 +1,5 @@
 import React from 'react';
-import {Row, Col, Icon} from 'antd';
+import {Icon} from 'antd';
 import request from '../../utils/request';
 import {connect} from 'react-redux';
 import {homeModule, informsState} from '../../redux/actions/createActions'
@@ -23,9 +23,9 @@ import {
   api_hot_word,
   api_carrier_pie,
   api_homepage_message,
-  api_save_widget
+  api_count_charts
 } from '../../services/api';
-import {formatOpinionCount, homeModuleList} from '../../utils/format';
+import {formatOpinionCount} from '../../utils/format';
 
 class NewHome extends React.Component {
   constructor() {
@@ -38,10 +38,12 @@ class NewHome extends React.Component {
       alldayWarningOpinion: [],
       weiboAll: [],
       weiboNegative: [],
-      opinionCountArr: [],
       hotWordData: [],
       mediaDistributionArr: [],
       homeMessage: [],
+      opinionCountArr:[],
+      opinionCount:{},
+      data:[],
       delMoudleList: {
         'todayOpinion': '今日舆情',
         'opinionTrend': '舆情走势',
@@ -53,28 +55,20 @@ class NewHome extends React.Component {
         'opinionCount': '舆情统计',
         'HotWord': '相关热词',
         'mediaDistribution': '媒体分布'
-      }
+      },
+      legend:[],
+      series:[]
     }
   }
 
-  componentWillMount() {
+  componentDidMount() {
     //首页模块
     request(api_homepage_message)
       .then(res => {
         if (res.data) {
           this.props.homeModule(res.data);
           this.setState({
-            homeMessage: res.data,
-            todayOpinion: res.data.todayOpinion,
-            mediaDistribution: res.data.mediaDistribution,
-            opinionTrend: res.data.opinionTrend,
-            newestOpinion: res.data.newestOpinion,
-            negativeOpinion: res.data.negativeOpinion,
-            newestWarningOpinion: res.data.newestWarningOpinion,
-            weiboOpinion: res.data.weiboOpinion,
-            topicOpinion: res.data.topicOpinion,
-            opinionCount: res.data.opinionCount,
-            HotWord: res.data.HotWord
+            homeMessage: res.data
           })
         }
             // 最新舆情
@@ -122,14 +116,23 @@ class NewHome extends React.Component {
                                 weiboNegative: res.data.negative
                               })
                             }
-
                             // 舆情统计
                             request(api_count_opinion)
-                              .then((res) => {
-                                const {opinionCountArr} = formatOpinionCount(res.data);
-                                this.setState({
-                                  opinionCountArr: opinionCountArr
-                                });
+                            .then( res => {
+                              if(res.data){
+                              this.setState({
+                                data: formatOpinionCount(res.data),
+                                opinionCount:res.data
+                                })
+                              }
+                                request(api_count_charts +`&data=${JSON.stringify(res.data.all)}`)
+                                .then(res => {
+                                  if(res.data.code === 1){
+                                    this.setState({
+                                      legend:res.data.legend,
+                                      series:res.data.series
+                                     })
+                                    }
                                     request(api_hot_word)
                                       .then(res => {
                                         this.setState({
@@ -144,328 +147,71 @@ class NewHome extends React.Component {
                                       })
                                   });
                               });
+                            })
                           });
                       });
                   });
-              });
+                });
   }
-
-  delHotWordBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.HotWord = 1;
-    this.setState({
-      HotWord: type,
-      homeMessage: homeMessage
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage)
-  }
-
-  delMediaDistributionBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.mediaDistribution = 1;
-    this.setState({
-      mediaDistribution: type,
-      homeMessage: homeMessage
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage)
-  }
-
-  delTodayBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.todayOpinion = 1;
-    this.setState({
-      todayOpinion: type,
-      homeMessage: homeMessage
-    })
-
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage);
-  }
-
-  delTrendBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.opinionTrend = 1;
-    this.setState({
-      opinionTrend: type,
-      homeMessage: homeMessage
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage);
-  }
-
-  delNewestBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.newestOpinion = 1;
-    this.setState({
-      newestOpinion: type,
-      homeMessage: homeMessage
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage);
-  }
-
-  delNegativeBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.negativeOpinion = 1;
-    this.setState({
-      negativeOpinion: type,
-      homeMessage: homeMessage
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage);
-  }
-
-  delNewestWarningBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.newestWarningOpinion = 1;
-    this.setState({
-      newestWarningOpinion: type,
-      homeMessage: homeMessage
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage);
-  }
-
-  delWeiboBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.weiboOpinion = 1;
-    this.setState({
-      weiboOpinion: type,
-      homeMessage: homeMessage
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage);
-  }
-
-  delTopicBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.topicOpinion = 1;
-    this.setState({
-      topicOpinion: type,
-      homeMessage: homeMessage
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage);
-  }
-
-  delCountBox(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage.opinionCount = 1;
-    this.setState({
-      opinionCount: type,
-      homeMessage: homeMessage
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage);
-  }
-
-  addModule(type) {
-    let homeMessage = this.state.homeMessage;
-    homeMessage[type] = 0;
-    this.setState({
-      homeMessage: homeMessage,
-      [type]: 0
-    })
-    request(api_save_widget, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded"
-      },
-      body: `widget=${JSON.stringify(homeMessage)}`
-    })
-    this.props.homeModule(homeMessage);
-  }
-
   informs() {
     this.props.informsState({data: false})
   }
-
+  changeChart(type){
+    request(api_count_charts +`&data=${JSON.stringify(this.state.opinionCount[type])}`)
+    .then(res => {
+      if(res.data.code === 1){
+        this.setState({
+          legend:res.data.legend,
+          series:res.data.series
+        })
+      }
+    })
+  }
   render() {
     const {opinionList,todayOpinionArr, alldayOpinion,todayWarningOpinion, alldayWarningOpinion, 
-      weiboAll, weiboNegative,opinionCountArr} = this.state;
-    const {userInfo, ModuleList} = this.props;
-    const moduleList = this.state.homeMessage.length !== 0 ? homeModuleList(this.state.homeMessage).map((item, index) =>
-      <li key={index}>{this.state.delMoudleList[item]}
-        <Icon type="plus-circle-o" className="addModule"
-              onClick={this.addModule.bind(this, item)}
-        ></Icon>
-      </li>) : '';
+      weiboAll, weiboNegative,opinionCountArr,homeMessage,hotWordData} = this.state;
+    const {userInfo,themeColor} = this.props;
+      const Notification = (state,size) => {
+          switch (state) {
+            case 'TodayOpinionBox':
+              return <TodayOpinionBox />;
+            case 'OpinionTrendBox':
+              return <OpinionTrendBox />;
+            case 'NewestOpinionBox':
+              return <NewestOpinionBox opinionList={opinionList}/>;
+            case 'NegativeOpinionBox':
+              return <NegativeOpinionBox  todayOpinion={todayOpinionArr} alldayOpinion={alldayOpinion}/>;
+            case 'NewestWarningOpinionBox':
+              return <NewestWarningOpinionBox todayOpinion={todayWarningOpinion} alldayOpinion={alldayWarningOpinion}/>;
+            case 'WeiboOpinionBox':
+              return <WeiboOpinionBox weiboAll={weiboAll} weiboNegative={weiboNegative}/>;
+            case 'OpinionCountBox':
+              return <OpinionCountBox data={opinionCountArr}/>;
+            case 'TopicOpinionBox':
+              return <TopicOpinionBox />;
+            case 'HotWordBox':
+              return <HotWordBox data={hotWordData}/>;
+            case 'MediaDistribution':
+              return <MediaDistribution data={this.state.mediaDistributionArr}/>;
+            default:
+              return null;
+          }
+        }  
     return (
-      <div>
+      <div style={{backgroundColor: themeColor.grounding.color}}>
         <div className="informs" style={userInfo.alerMsg === '' ? {display: 'none'} : {display: 'flex'}}>
           <span> {userInfo.alerMsg}</span>
           <Icon type="close" onClick={this.informs.bind(this)} style={{color: '#fff'}}/>
         </div>
-        <div className="home-pages"
-             style={this.props.type !== undefined ? {'backgroundColor': '#ffffff'} : {'backgroundColor': '#e4ebf7'}}>
-          <div className="home-layout"
-               style={this.props.type !== undefined ? {display: 'block', width: '8%'} : {display: 'none'}}>
-            <div className="layout-list">
-              <div>首页</div>
-              <ul>
-                {moduleList}
-              </ul>
-            </div>
+        <div className="home-pages" >
+          <div className="container" style={this.props.type !== undefined ? {width: '92%'} : {width: '100%',display:'flex',flexWrap:'wrap'}}>
+         { homeMessage.map((item,index) => {
+            return <div key={index} style={{width:`${item.defaultSize-2}%`,margin:'1%'}}>{Notification(item.name,item.defaultSize)}</div>
+          })
+         }
           </div>
-          <div className="container" style={this.props.type !== undefined ? {width: '92%'} : {width: '100%'}}>
-            <Row gutter={16} className="row"
-                 style={ModuleList.todayOpinion === 1 ? {display: 'none'} : {display: 'block'}}
-            >
-              <Col span={24}
-              >
-                <TodayOpinionBox status={this.props.type !== undefined ? 'setting' : ''}
-                                 delTodayBox={this.delTodayBox.bind(this)}                                
-                />
-              </Col>
-            </Row>
-            <Row gutter={16} className="row"
-            >
-              <Col span={24}>
-                {ModuleList.opinionTrend === 1 ? '' : <OpinionTrendBox
-                  status={this.props.type !== undefined ? 'setting' : ''}
-                  delTrendBox={this.delTrendBox.bind(this)}
-                />}
-
-              </Col>
-            </Row>
-            <Row gutter={16} className="row">
-              <Col span={12}
-                   style={ModuleList.newestWarningOpinion === 1 ? {display: 'none'} : {display: 'block'}}
-              >
-                <NewestWarningOpinionBox
-                  todayOpinion={todayWarningOpinion}
-                  alldayOpinion={alldayWarningOpinion}
-                  status={this.props.type !== undefined ? 'setting' : ''}
-                  delNewestWarningBox={this.delNewestWarningBox.bind(this)}
-                />
-              </Col>
-              <Col span={12}
-                   style={ModuleList.negativeOpinion === 1 ? {display: 'none'} : {display: 'block'}}
-              >
-                <NegativeOpinionBox
-                  todayOpinion={todayOpinionArr}
-                  alldayOpinion={alldayOpinion}
-                  status={this.props.type !== undefined ? 'setting' : ''}
-                  delNegativeBox={this.delNegativeBox.bind(this)}
-                />
-              </Col>
-              <Col span={12}
-                   style={ModuleList.newestOpinion === 1 ? {display: 'none'} : {display: 'block'}}
-              >
-                <NewestOpinionBox opinionList={opinionList}
-                                  status={this.props.type !== undefined ? 'setting' : ''}
-                                  delNewestBox={this.delNewestBox.bind(this)}
-                />
-              </Col>
-              <Col span={12}
-                   style={ModuleList.weiboOpinion === 1 ? {display: 'none'} : {display: 'block'}}
-              >
-                <WeiboOpinionBox
-                  weiboAll={weiboAll}
-                  weiboNegative={weiboNegative}
-                  status={this.props.type !== undefined ? 'setting' : ''}
-                  delWeiboBox={this.delWeiboBox.bind(this)}
-                />
-              </Col>
-              <Col span={12}
-                   style={ModuleList.topicOpinion === 1 ? {display: 'none'} : {display: 'block'}}
-              >
-                <TopicOpinionBox status={this.props.type !== undefined ? 'setting' : ''}
-                                 delTopicBox={this.delTopicBox.bind(this)}
-                />
-              </Col>
-              <Col span={12}
-                   style={ModuleList.opinionCount === 1 ? {display: 'none'} : {display: 'block'}}
-              >
-                <OpinionCountBox data={opinionCountArr}
-                                 status={this.props.type !== undefined ? 'setting' : ''}
-                                 delCountBox={this.delCountBox.bind(this)}
-                />
-              </Col>
-
-              <Col span={12}
-                   style={ModuleList.HotWord === 1 ? {display: 'none'} : {display: 'block'}}
-              >
-                <HotWordBox data={this.state.hotWordData}
-                            status={this.props.type !== undefined ? 'setting' : ''}
-                            delHotWordBox={this.delHotWordBox.bind(this)}
-                />
-              </Col>
-              <Col span={12}
-              >
-                {ModuleList.mediaDistribution === 1 ? '' : <MediaDistribution
-                  data={this.state.mediaDistributionArr}
-                  status={this.props.type !== undefined ? 'setting' : ''}
-                  delMediaDistributionBox={this.delMediaDistributionBox.bind(this)}
-                />}
-
-              </Col>
-            </Row>
           </div>
         </div>
-      </div>
     )
   }
 }
@@ -474,7 +220,8 @@ const mapStateToProps = state => {
   return {
     ModuleList: state.homeModuleReducer,
     userInfo: state.userFetchSuccess,
-    informsstate: state.informsstate.data
+    informsstate: state.informsstate.data,
+    themeColor: state.changeThemeReducer
   }
 };
 const mapDispatchToProps = dispatch => {
