@@ -1,15 +1,16 @@
 import React from 'react';
 import alone from './../../assets/icon-img/alone.png';
+import blank from './../../assets/icon-img/blank.png';
 import './BlankPage.less';
-//import default from '../../../node_modules/_antd@2.13.10@antd/lib/notification';
 class Blank extends React.Component{
       render(){          
            const desc=this.props.desc;
+           const status = this.props.status;
            return(
              <div className="blank-page-wrapper">
             <div className="img-wrapper">
                 <img
-                    src={alone}
+                    src={status?alone:blank}
                     alt="blank img"
                     className="blank-img"
                 />
