@@ -1,10 +1,9 @@
 import React from 'react';
 import {Icon} from 'antd';
-import { LIGHT, DARK,} from '../../utils/colors';
+import { LIGHT, DARKER,BLUES,GRAY} from '../../utils/colors';
 import {changeTheme} from '../../redux/actions/actions';
 import {connect} from 'react-redux';
 import './ChangeTheme.less';
-
 class ChangeTheme extends React.Component {
     constructor(props) {
         super(props);
@@ -14,42 +13,38 @@ class ChangeTheme extends React.Component {
             theme: [
                 {
                     topColor: {
-                        backgroundColor: LIGHT
+                        backgroundColor: BLUES
                     },
                     bottomColor: {
-                        backgroundColor: DARK
+                        backgroundColor: '#fff'
                     },
                     textColor:{
-                         color:'#000000'
+                        color:'#000'
+                    },
+                    grounding:{
+                        color:'#E4EBF7'
+                    },
+                    borderColor:{
+                        color:'#e4ebf7'
                     }
                 },
                 {
                     topColor: {
-                        backgroundColor: DARK
+                        backgroundColor: LIGHT
                     },
                     bottomColor: {
-                        backgroundColor: DARK
+                        backgroundColor: DARKER
                     },
                     textColor:{
-                        color:'rgba(255, 255, 255, 0.67)'
+                        color:'#b5b5b5'
+                    },
+                    grounding:{
+                        color:'#181b2b'
+                    },
+                    borderColor:{
+                        color:'#35394f'
                     }
-                },
-                // {
-                //     topColor: {
-                //         backgroundColor: BLUE
-                //     },
-                //     bottomColor: {
-                //         backgroundColor: LIGHT
-                //     }
-                // },
-                // {
-                //     topColor: {
-                //         backgroundColor: RED
-                //     },
-                //     bottomColor: {
-                //         backgroundColor: LIGHT
-                //     }
-                // }
+                }
             ]
         }
     }
@@ -60,7 +55,8 @@ class ChangeTheme extends React.Component {
         });
         this.props.changeTheme(item);
     }
-    render() {
+    render() { 
+        const {themeColor} = this.props;
         const Themes = this.state.theme.map((item, index) =>
             <div className="theme-item" key={index}
                  onClick={this.showItem.bind(this, index, item)}
@@ -69,7 +65,7 @@ class ChangeTheme extends React.Component {
                 <div className="bottom">
                     <div className="left" style={item.bottomColor}></div>
                     <div className="right">
-                        <Icon type="check-square" className="check" style={this.state.index === index ? {display: 'block'} : {display: 'none'}}/>
+                        <Icon type="check-square" className="check" style={themeColor.topColor.backgroundColor === item.topColor.backgroundColor ? {display: 'block'} : {display: 'none'}}/>
                     </div>
                 </div>
             </div>
@@ -84,6 +80,11 @@ class ChangeTheme extends React.Component {
         )
     }
 }
+const mapStateToProps = state => {
+    return {
+        themeColor: state.changeThemeReducer
+    }
+};
 const mapDispatchToProps = dispatch => {
     return {
         changeTheme: req => {
@@ -91,4 +92,4 @@ const mapDispatchToProps = dispatch => {
         }
     }
 };
-export default connect(null, mapDispatchToProps)(ChangeTheme);
+export default connect(mapStateToProps, mapDispatchToProps)(ChangeTheme);
