@@ -1,44 +1,38 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import {Layout, Menu,Button} from 'antd';
-import {Route, Switch, Link} from 'react-router-dom';
+import {Layout,Button} from 'antd';
+import {Route, Switch} from 'react-router-dom';
 import './Index.less';
 import {exportSkip} from '../../redux/actions/createActions';
-import logo from '../../assets/img/logo.png';
 import CRcode from '../LoginPage/crcode.jpg';
-import {api_get_channel} from '../../services/api';
-import Iconfont from '../../components/IconFont'
-import request from '../../utils/request';
-import {urlTokey} from '../../utils/format';
 import AsyncComponent from '../../components/AsyncComponent/AsyncComponent'
 const NewHome = AsyncComponent(() => import('../NewHome'))
 const TopicReportList = AsyncComponent(() => import('../TopicReportList/TopicReportList'))
-const SortedAdd = AsyncComponent(() => import('../SortedOpinion/SortedAdd'))
-const TopicAdd = AsyncComponent(() => import('../TopicOpinion/TopicAdd/TopicAdd'))
-const SortedOpinion = AsyncComponent(() => import('../../routes/SortedOpinion'))
-const ExcludeSetting = AsyncComponent(() => import('../SystemSetting/ExcludeSetting/ExcludeSetting'))
-const WarnSetting = AsyncComponent(() => import('../SystemSetting/WarnSetting/WarnSetting'))
-const NoticeSetting = AsyncComponent(() => import('../SystemSetting/NoticeSetting/NoticeSetting'))
-const TopicEditOpinionDetail = AsyncComponent(() => import('../SystemSetting/TopicEditOpinionDetail/TopicEditOpinionDetail'))
-const EditOpinionDetail = AsyncComponent(() => import('../../components/EditOpinionDetail/EditOpinionDetail'))
-const HistoryOpinion = AsyncComponent(() => import('../HistoryOpinion/HistoryOpinion'))
-const CollectionOpinion = AsyncComponent(() => import('../CollectionOpinion/CollectionOpinion'))
-const MaterialOpinion = AsyncComponent(() => import('../MaterialOpinion/MaterialOpinion'))
-const ReportOpinionDetail = AsyncComponent(() => import('../ReportOpinion/ReportOpinionDetail'))
-const ReportOpinion = AsyncComponent(() => import('../ReportOpinion/ReportOpinion'))
-const TopicOpinion = AsyncComponent(() => import('../TopicOpinion/TopicOpinion'))
 const AllOpinion = AsyncComponent(() => import('../AllOpinion/AllOpinion'))
 const TrendFeeling = AsyncComponent(() => import('../TopicOpinion/TrendFeeling/TrendFeeling'))
-const BigScreen = AsyncComponent(() => import('../BigScreen/BigScreen'))
 const ZHeader = AsyncComponent(() => import('../../components/ZHeader/Zheader'))
-const ReportTemplate = AsyncComponent(() => import('../ReportTemplate/ReportTemplate'))
-const ChooseTemplate = AsyncComponent(() => import('../ChooseTemplate/ChooseTemplate'))
-const CustomHome= AsyncComponent(() => import('../CustomHome/CustomHome'))
-const MyReport= AsyncComponent(() => import('../MyReport/MyReport'))
-const Drag= AsyncComponent(() => import('../Drag/Drag'))
-const {Header, Sider, Content} = Layout;
-const SubMenu = Menu.SubMenu;
-
+const Evidence = AsyncComponent(() => import('../Evidence/Evidence'))
+const UpReport = AsyncComponent(() => import('../UpReport/UpReport'))
+const Guide = AsyncComponent(() => import('../Guide/Guide'))
+const Situational = AsyncComponent(() => import('../Situational/Situational'))
+const AppCenter= AsyncComponent(() => import('../AppCenter/AppCenter'))
+const Guard= AsyncComponent(() => import('../Guard/Guard'))
+const MonitoringWarning = AsyncComponent(() => import('../MonitoringWarning/MonitoringWarning'))
+const SecurityGovernance = AsyncComponent(() => import('../SecurityGovernance/SecurityGovernance'))
+const SpaceExploration = AsyncComponent(() => import('../SpaceExploration/SpaceExploration'))
+const SafetyProtection = AsyncComponent(() => import('../SafetyProtection/SafetyProtection'))
+const AlertNotifications = AsyncComponent(() => import('../AlertNotifications/AlertNotifications'))
+const FmEngine = AsyncComponent(() => import('../FmEngine/FmEngine'))
+const FmLgScreen = AsyncComponent(() => import('../FmLgScreen/FmLgScreen'))
+const Disposal = AsyncComponent(() => import('../Disposal/Disposal'))
+const PartyBuilding = AsyncComponent(() => import('../PartyBuilding/PartyBuilding'))
+const IndustryInformation = AsyncComponent(() => import('../IndustryInformation/IndustryInformation'))
+const CompetitiveIntelligence = AsyncComponent(() => import('../CompetitiveIntelligence/CompetitiveIntelligence'))
+const WeChatFence = AsyncComponent(() => import('../WeChatFence/WeChatFence'))
+const CloudDisk = AsyncComponent(() => import('../CloudDisk/CloudDisk'))
+const CloudPlatform = AsyncComponent(() => import('../CloudPlatform/CloudPlatform'))
+const DeadwoodCreep = AsyncComponent(() => import('../DeadwoodCreep/DeadwoodCreep'))
+const { Content} = Layout;
 class Index extends React.Component {
   constructor() {
     super();
@@ -71,17 +65,6 @@ class Index extends React.Component {
     })
   }
  }
-  componentWillMount() {
-    request(api_get_channel)
-      .then(res => {
-        if (res.data.code === 1) {
-          this.setState({
-            channelList: res.data.channelList
-          })
-        }
-      })
-  }
-
   goBackIndex() {
       if(this.state.collapsed){
       this.setState({
@@ -135,148 +118,46 @@ class Index extends React.Component {
   backTop() {
     document.querySelector('.main').scrollTop = 0;
   }
+  onEnter = () => {
+    console.log('进来了')
+  }
   render() {
     // 统计报告-舆情报告   
-    const {themeColor} = this.props;
-    let menuList = [];
-    this.state.channelList.map((item, index) => {
-      if (item.channelurl === '/reportopinion/list') {
-        menuList.push(<SubMenu
-          key={item.key}
-          title={<Link to="/historyopinion"><span>
-            <i className="anticon"><Iconfont type={item.type} style={{fontSize: '16px'}}/></i>
-            <span
-              style={{fontSize: '16px'}}>舆情报告</span>
-            </span></Link>}>
-          <Menu.Item key="reportopinion" style={{fontSize: '16px'}}>
-            <Link to="/reportopinion/list">
-              <span>简报列表</span>
-            </Link>
-          </Menu.Item>
-          <Menu.Item key="materiaopinion" style={{fontSize: '16px'}}>
-            <Link to="/materiaopinion">
-              <span>素材库</span>
-            </Link>
-          </Menu.Item>
-          <Menu.Item key="collectionopinion" style={{fontSize: '16px'}}>
-            <Link to="/collectionopinion">
-              <span>我的收藏</span>
-            </Link>
-          </Menu.Item>
-          <Menu.Item key="historyopinion" style={{fontSize: '16px'}}>
-            <Link to="/historyopinion">
-              <span>我的报告库</span>
-            </Link>
-          </Menu.Item>
-        </SubMenu>)
-      } else if (item.channelurl === '../systemMan/systemManDo?action=userList') {
-        menuList.push(<SubMenu key={item.key} 
-                               title={<Link to="/noticesetting"><span><i className="anticon"><Iconfont type={item.type}
-                                 style={{fontSize: '16px'}}/></i><span
-                                 style={{fontSize: '16px'}}>系统设置</span></span> </Link>}>
-          <Menu.Item key="noticesetting" style={{fontSize: '16px'}}>
-            <Link to="/noticesetting">
-              <span>通知设置</span>
-            </Link>
-          </Menu.Item>
-          
-          <Menu.Item key="warnsetting" style={{fontSize: '16px'}}>
-            <Link to="/warnsetting">
-              <span>预警设置</span>
-            </Link>
-          </Menu.Item>
-          <Menu.Item key="excludesetting" style={{fontSize: '16px'}}>
-            <Link to="/excludesetting">
-              <span>排除停用</span>
-            </Link>
-          </Menu.Item>
-          <Menu.Item key="publicopinion" style={{fontSize: '16px'}}>
-            <Link to="/publicopinion">
-              <span>舆情录入</span>
-            </Link>
-          </Menu.Item>
-        </SubMenu>)
-      } else {
-        menuList.push(<Menu.Item key={item.key} style={{fontSize: '16px'}}>
-          {item.channelurl.indexOf('http') !== -1 ?
-            <a href={item.channelurl} target="blank">
-              <i className="anticon"><Iconfont type={item.type} style={{fontSize: '16px'}}/></i>
-              <span>{item.channelname}</span>
-            </a> : <Link to={item.channelurl}>
-            <i className="anticon"><Iconfont type={item.type} style={{fontSize: '16px'}}/></i>
-              <span>{item.channelname}</span>
-            </Link>
-          }
-        </Menu.Item>)
-      }
-      return 3
-    })
     return (
       <div className="root-container">
-        <Layout className="layout">
-          <Sider
-            className="sider siders"
-            trigger={null}
-            collapsible
-            style={{backgroundColor: '#fff'}}
-            collapsed={this.state.collapsed && this.state.flag}
-            onMouseEnter={this.mouseEnterToggle} 
-            onMouseLeave={this.mouseLeaveToggle}
-          >
-            <div>
-              <div className="logo-wrapper">
-                <img src={logo} alt="logo" className="logo" onClick={this.goBackIndex.bind(this)}/>
-              </div>
-              <div className="trigger-wrapper" onClick={this.toggle} >
-                <i className="fa fa-bars" aria-hidden="true" style={{fontSize: '14px', color: '#5a8bff'}}/>
-              </div>
-            </div>
-            <Menu
-              defaultSelectedKeys={['1']}
-              mode="inline"
-              theme="dark"
-              // inlineCollapsed={this.state.collapsed && this.state.flag}
-              style={ {backgroundColor: '#fff', overflow: 'auto',maxHeight: '600px'}}
-              className="selectMenu"
-              selectedKeys={[urlTokey()]}
-            >
-              {menuList}
-            </Menu>
-          </Sider>
+       <ZHeader/> 
           <Layout className="right-layout">
-            <Header className="header" style={{backgroundColor: themeColor.topColor.backgroundColor, height: '60px'}}>
-              <ZHeader/>
-            </Header>
             <Content className="main" ref={(main) => {
               this.main = main
             }}>
               <Switch>
                 <Route path="/trendfeeling" component={TrendFeeling}/>
-                <Route path="/bigscreen" component={BigScreen}/>
                 <Route path="/allopinion" component={AllOpinion}/>
-                <Route path="/topic" component={TopicOpinion}/>
-                <Route path="/test" component={EditOpinionDetail}/>
-                <Route path="/reportopinion/list" component={ReportOpinion}/>
-                <Route path="/reportopinion/detail" component={ReportOpinionDetail}/>
-                <Route path="/materiaopinion" component={MaterialOpinion}/>
-                <Route path="/collectionopinion" component={CollectionOpinion}/>
-                <Route path="/historyopinion" component={HistoryOpinion}/>
                 <Route path="/home" exact component={NewHome}/>
-                <Route path="/noticesetting" component={NoticeSetting}/>
-                <Route path="/warnsetting" component={WarnSetting}/>
-                <Route path="/excludesetting" component={ExcludeSetting}/>
-                <Route path="/publicopinion" component={TopicEditOpinionDetail}/>
-                <Route path="/sortedopinion" component={SortedOpinion}/>
-                <Route path="/topic/addtopic" component={TopicAdd}/>
-                <Route path="/sortedopinion/addrule" component={SortedAdd}/>
                 <Route path="/topicreportlist" component={TopicReportList}/>
-                <Route path="/reporttemplate" component={ReportTemplate}/>
-                <Route path="/choosetemplate" component={ChooseTemplate}/>
-                <Route path="/customhome" component={CustomHome}/>
-                <Route path="/myreport" component={MyReport}/>
-                <Route path="/drag" component={Drag}/>
+                <Route path="/evidence" component={Evidence}/>
+                <Route path="/upreport" component={UpReport}/>
+                <Route path="/guide" component={Guide}/>
+                <Route path="/situational" component={Situational}/>
+                <Route path="/appcenter" component={AppCenter}  onEnter={this.onEnter}/> 
+                <Route path="/Guard" component={Guard}/>
+                <Route path="/monitoringwarning" component={MonitoringWarning}/>
+                <Route path="/securitygovernance" component={SecurityGovernance}/>
+                <Route path="/spaceexploration" component={SpaceExploration}/>
+                <Route path="/safetyprotection" component={SafetyProtection}/>
+                <Route path="/alertnotifications" component={AlertNotifications}/>
+                <Route path="/fmengine" component={FmEngine}/>
+                <Route path="/fmlgscreen" component={FmLgScreen}/>
+                <Route path="/disposal" component={Disposal}/>
+                <Route path="/partybuilding" component={PartyBuilding}/>
+                <Route path="/industryinformation" component={IndustryInformation}/>
+                <Route path="/competitiveintelligence" component={CompetitiveIntelligence}/>
+                <Route path="/wechatfence" component={WeChatFence}/>
+                <Route path="/clouddisk" component={CloudDisk}/>
+                <Route path="/cloudplatform" component={CloudPlatform}/>
+                <Route path="/deadwoodcreep" component={DeadwoodCreep}/>
               </Switch>
-              <div className="suspensionBox">
+              <div className="suspensionBox" style={{display:"none"}}>
                 <div>
                   <ul className="suspension">
                     <li
@@ -335,7 +216,6 @@ class Index extends React.Component {
               </div>
             </Content>
           </Layout>
-        </Layout>
       </div>
     )
   }
