@@ -1,4 +1,5 @@
 import React from 'react';
+import {connect} from 'react-redux';
 import './index.less';
 import { Form, Icon, Input, Button, Checkbox, Col, Row, message} from 'antd';
 import zhiyunImg from './zhiyun.png';
@@ -9,7 +10,8 @@ import {history} from '../../utils/history';
 import CRCode from './crcode.jpg';
 import IosApp from './iosapp.jpg';
 import {setItem,getPasswordItem,setPasswordItem} from '../../utils/localStorage';
-import {LIGHT,DARK} from '../../utils/colors';
+import {changeTheme} from '../../redux/actions/actions';
+import {LIGHT,DARKER} from '../../utils/colors';
 
 const FormItem = Form.Item;
 
@@ -59,17 +61,40 @@ class LoginPage extends React.Component {
                   checked:true
                })
            }
-           setItem('theme',{
+           let colors ={
+            // topColor: {
+            //     backgroundColor: BLUES
+            // },
+            // bottomColor: {
+            //     backgroundColor: '#fff'
+            // },
+            // textColor:{
+            //      color:'#000'
+            // },
+            // grounding:{
+            //     color:'#E4EBF7'
+            // },
+            // borderColor:{
+            //     color:'#e4ebf7'
+            // },
             topColor: {
                 backgroundColor: LIGHT
             },
             bottomColor: {
-                backgroundColor: DARK
+                backgroundColor: DARKER
             },
             textColor:{
-                 color:'#fff'
+                color:'#b5b5b5'
+            },
+            grounding:{
+                color:'#181b2b'
+            },
+            borderColor:{
+                color:'#35394f'
             }
-        });
+           };
+           setItem('theme',colors);
+           this.props.changeTheme(colors)
     }
     handleSubmit (e){
         e.preventDefault();
@@ -220,5 +245,11 @@ class LoginPage extends React.Component {
         )
     }
 }
-
-export default Form.create()(LoginPage);
+const mapDispatchToProps = dispatch => {
+    return {
+        changeTheme: req => {
+            dispatch(changeTheme(req));
+        }
+    }
+};
+export default connect(null, mapDispatchToProps)(Form.create()(LoginPage));;
