@@ -1,17 +1,18 @@
 import React from 'react';
 import './MediaDistribution.less';
+import {connect} from 'react-redux';
 import IconFont from '../../components/IconFont';
 import ReactEchartsCore from 'echarts-for-react/lib/core';
 import echarts from 'echarts/lib/echarts';
 import 'echarts/lib/chart/pie';
 import {Icon} from 'antd';
-import {GRAY,BLACK,BLUES} from '../../utils/colors';
+import {BLUES} from '../../utils/colors';
 class MediaDistribution extends React.Component{
     delmediaDistributionBox(){
         this.props.delMediaDistributionBox(1);
     }
     render(){
-        const {data} = this.props;
+        const {data,themeColor} = this.props;
 				const mediaOption= {
 					tooltip: {
 						trigger: 'item',
@@ -36,12 +37,12 @@ class MediaDistribution extends React.Component{
 					]
         };     
         return (
-            <div className="mediaDistribution-box">
+            <div className="mediaDistribution-box" style={{background:themeColor.bottomColor.backgroundColor}}>
             <div className="container">
-            <div className="top" style={{background:GRAY}}>
+            <div className="top" style={{borderBottom: `1px solid ${themeColor.borderColor.color}`}}>
                 <div className="title">
                     <IconFont type="icon-fenbutu" style={{fontSize: '22px',color:BLUES}}/>
-                    <span className="txt" style={{color:BLACK}}>媒体分布</span>
+                    <span className="txt" style={{color:themeColor.textColor.color}}>媒体分布</span>
                 </div>
                 <div className="more">
                         <Icon type="close-circle" className="delModule"
@@ -61,4 +62,9 @@ class MediaDistribution extends React.Component{
         )
     }
 }
-export default MediaDistribution;
+const mapStateToProps = state => {
+    return {
+      themeColor: state.changeThemeReducer
+    }
+  };
+export default  connect(mapStateToProps, null)(MediaDistribution);
