@@ -57,7 +57,7 @@ class SortedAdd extends React.Component {
                         message.success(res.data.msg);
                         getSortedMenuRequested();
                         history.push({
-                            pathname: '/sortedopinion/list',
+                            pathname: '/allopinion/sortedopinion/list',
                             search:`?clfid=${clfId}`
                         })
                     } else {
@@ -105,7 +105,7 @@ class SortedAdd extends React.Component {
         const clfId = this.props.clfId;
         getSortedMenuRequested();
         history.push({
-            pathname: '/sortedopinion/list',
+            pathname: '/allopinion/sortedopinion/list',
             search:`?clfid=${clfId}`
         })
     }
