@@ -67,6 +67,7 @@ class SortedSetting extends React.Component {
           message.error('话题名称请不要为空');
           return;
         }
+        console.log(rules)
         request(api_sorted_rule_edit, {
           method: 'POST',
           headers: {
@@ -78,7 +79,7 @@ class SortedSetting extends React.Component {
             getSortedMenuRequested();
             message.success(res.data.msg)
             history.push({
-              pathname: '/sortedopinion/list',
+              pathname: '/allopinion/sortedopinion/list',
               search: `?clfid=${clfId}`
             })
           }
@@ -93,7 +94,7 @@ class SortedSetting extends React.Component {
     const clfId = this.props.clfId;
     getSortedMenuRequested();
     history.push({
-      pathname: '/sortedopinion/list',
+      pathname: '/allopinion/sortedopinion/list',
       search: `?clfid=${clfId}`
     })
   }
@@ -114,8 +115,8 @@ class SortedSetting extends React.Component {
     this.setState({
       num3: this.state.num3.concat([{
         "rule1": "", "rulecode1": "", "id": "", "rule2": "",
-        "rulecode2": "", "rule3": "", "rulecode3": "", "rule4": "","rule":"",
-        "rulecode4": ""
+        "rulecode2": "", "rule3": "", "rulecode3": "", "rule4": "",
+        "rulecode4": "","rule":""
       }])
     });
   }
@@ -193,12 +194,12 @@ class SortedSetting extends React.Component {
 
   onInputConent(e) {
     let rule = [];
+    
     if (e.length > 0) {
       for (let i in e) {
         rule.push({});
         rule[i]['rule'] = e[i].rule;
         rule[i]['id'] = e[i].id;
-
       }
     }
     this.setState({
