@@ -1,7 +1,7 @@
 export const BLUE = '#108EE9';
 export const DARK = '#2d324f';
-export const DARKER = '#2c2c2c';
-export const LIGHT = '#5a8bff';
+export const DARKER = '#25283a';
+export const LIGHT = '#35394f';
 export const RED = '#F01414';
 export const GRAY = '#f0f2fb';
 export const BLACK = '#000';
